Ignore test on windows & fix checkstyle
